Prevent the #media_url from being overridden
There is a race between the upload js and the js to set the `#media_url` from the add media form (`#url`).  We should only ever set the value of `#media_url` with `#url` IFF it is NOT an upload.  Otherwise, we could upload the video, set the `#media_url` with the uploaded url and then the rest of the script would execute and set the `#media_url` to null, because a `#url` was never provided.