Improve accessibility of upload
Add a label to the file upload input and change the link to a div

plupload does not provide a label for inputs.  The input is in the DOM, not hidden, but placed behind the upload `container`, so a label is still required.  The input is accessible in the tab order of the document, but so was the href.  This could have been confusing, so I changed the href to a div and the same functionality exists.