Only don't reload the map on `#get_location` click
`#geo_location` is the fieldset containing the map.  If you click the map, the parent onClick also fires, thus re-loading the map.

It looks like this `#geo_location` click handler is meant to only initialize the google map once the section is shown.  It is probably meant to reduce the work that the page has to do on initial load.