Show all languages and only show captions if they exist
This improves captions by adding all language tracks.  It also does not load any tracks if none exist, which will prevent the never ending 'english loading...' text from appearing.

The only downfall is a potential longer page load due to a query to the amara API for each load.  This will be addressed in another commit.