Fix case where video dimensions were not being saved
Video dimensions were not being saved because they were being computed based on the thumnail (probably because it is easier to figure out the dimensions of an image than a video). Besides the fact that the thumbnail might not match the video size, the thumbnail doesn't exist when the dimensions are first cached. So... reference the video instead of the thumbnail.