Fix HLS captions toggle button
The HLS library will add a metadata text track and then remove it after the video loads. This changes the indexes of the text tracks, and for whatever reason this change does not cause `textTracks.on("change")` to be called. Thus, depending on an index value was not safe and resulted in an error under some circumstances.